AOS-3404 Move fuzzy.go to pkg auroraconfig
