+++ pkg/auroraconfig/fuzzy.go
@@ -1,4 +1,4 @@
-package fuzzy
+package auroraconfig
 
 import (
 	"path/filepath"
