+++ pkg/auroraconfig/fuzzy_test.go
@@ -1,4 +1,4 @@
-package fuzzy
+package auroraconfig
 
 import (
 	"testing"
